+ more 64-bit stuff

+++ src/Mod/Mesh/Gui/ViewProvider.cpp
@@ -180,7 +180,7 @@ ViewProviderMesh::ViewProviderMesh() : m_bEdit(false)
   unsigned long setting = hGrp->GetUnsigned("MeshColor", current);
   if ( current != setting )
   {
-    color.setPackedValue(setting);
+    color.setPackedValue((uint32_t)setting);
     ShapeColor.setValue(color);
   }
 }
